% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.r
\name{gini}
\alias{gini}
\title{gini}
\usage{
gini(x, y)
}
\arguments{
\item{\code{x}}{: within strata mean per tow}

\item{\code{y}}{: strata areas in trawlable units}
}
\value{
estimate of gini
}
\description{
gini index of patchniness from a stratified survey (Myers and Cadigan 1995)
}
\examples{
y=c(2454,	656,	2374,	230,	1221,	619,	172,	2980,	360,	3245,	1016,	2569,	424,	390,	454)
x=c(0,	0,	0,	0,	0,	0.25,	1.81,	2.52,	2.6,	3.68,	7.58,	10.07,	13.33,	13.33,	24.99)
gini(x,y)
#0.453 same as example in http://www.nefsc.noaa.gov/publications/crd/pdfs/crd9605f.pdf
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}

